separate kube and openshift admission
