+++ pkg/cmd/openshift-apiserver/openshiftadmission/plugin_initializer.go
@@ -1,4 +1,4 @@
-package admission
+package openshiftadmission
 
 import (
 	"fmt"
